refactor: Moved asp.net core project into a separate folder "AspCoreServer"
The Angular2Spa folder was a mess with the node/spa code and the .net core code sharing the same folder.
This refactoring moves the core project to a separate folder and redirects the webpack dist folders to route to the correct places in the .net core folder.
