+++ src/Angular2Spa/webpack.config.test.vendor.js
@@ -37,7 +37,7 @@ module.exports = {
         ]
     },
     output: {
-        path: path.join(__dirname, 'wwwroot', 'dist'),
+        path: path.join(__dirname, '../AspCoreServer/wwwroot', 'dist'),
         filename: '[name].js',
         library: '[name]_[hash]',
     },
@@ -50,7 +50,7 @@ module.exports = {
         extractCSS,
         //new webpack.ProvidePlugin({ $: 'jquery', jQuery: 'jquery' }), // Maps these identifiers to the jQuery package (because Bootstrap expects it to be a global variable)
         new webpack.DllPlugin({
-            path: path.join(__dirname, 'wwwroot', 'dist', '[name]-manifest.json'),
+            path: path.join(__dirname, '../AspCoreServer/wwwroot', 'dist', '[name]-manifest.json'),
             name: '[name]_[hash]'
         })
     ].concat(isDevelopment ? [] : [
