+++ src/AspCoreServer/Dockerfile
