+++ src/AspCoreServer/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-@using Angular2Spa
+@using AspCoreServer
 @addTagHelper "*, Microsoft.AspNetCore.Mvc.TagHelpers"
 @addTagHelper "*, Microsoft.AspNetCore.SpaServices"
