+++ src/AspCoreServer/app.json
