+++ src/AspCoreServer/project.json
@@ -76,7 +76,7 @@
     ]
   },
   "tooling": {
-    "defaultNamespace": "Angular2Spa"
+    "defaultNamespace": "AspCoreServer"
   },
   "tools": {
     "Microsoft.AspNetCore.Razor.Tools": "1.0.0-preview2-final",
