+++ test/AspCoreServer.Test/AspCoreServer.Test.xproj
@@ -7,7 +7,7 @@
   <Import Project="$(VSToolsPath)\DotNet\Microsoft.DotNet.Props" Condition="'$(VSToolsPath)' != ''" />
   <PropertyGroup Label="Globals">
     <ProjectGuid>d7677314-5b09-4c9a-a695-f046ace5a97f</ProjectGuid>
-    <RootNamespace>Angular2Spa.Test</RootNamespace>
+    <RootNamespace>AspCoreServer.Test</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">.\obj</BaseIntermediateOutputPath>
     <OutputPath Condition="'$(OutputPath)'=='' ">.\bin\</OutputPath>
   </PropertyGroup>
