+++ test/AspCoreServer.Test/project.json
@@ -4,7 +4,7 @@
     "debugType": "portable"
   },
   "dependencies": {
-    "Angular2Spa": "1.0.0-*",
+    "AspCoreServer": "1.0.0-*",
     "dotnet-test-xunit": "2.2.0-preview2-build1029",
     "Microsoft.DotNet.InternalAbstractions": "1.0.0",
     "System.Runtime.Serialization.Primitives": "4.3.0",
